Readme.md
# scripts


The following project is made in Bash and is used for IT Infrastructure automation in Wikipedia.
You need Redmine for this, as right now it is only Redmine based Wikipedia.

The Command script is called "files.sh" This will start all the proccesses required for this project. 
This one must be used in crontab to automate it.

"**Set -x**" can be removed from every script, i have added it for debugging.

SCRIPT OVERVIEW and what they do exactly.

 **files.sh**
 
The following script has many variables. You have the location for every script that will be executed.
You also have variables for each Physical Host.

*SendFiles_Install*  
Function is used to send "proginst.sh" into all of the Physical Machines.
It is also used to send "maht.sh" and "memmaht.sh" to each Physical Machine.

*SendFiles_Physical*  
Function is used to send "servers.sh" script to each Physical Machine and to execute it.
The following function also sends back the data to the Commander Machine (where the script was executed) that "servers.sh" has collected.

*SendFiles_Virtual*  
Function is used to send "virtualservers.sh" script to each Virtual Machine and to execute it.
Data is also sent back to the Commander Machine.

*SendFiles_Redmine*  
Function is used to send the final "Alltogether.xml" file to the Redmine server.
This also executes a script "rmine.sh" 

------------------------------------------------------------------------
Now we describe what each script does.

You have Upper.txt and Lower.txt 
These 2 files are used to hold Wikipedia xml headers.
Upper.txt holds the upper header and Lower.txt holds the lower header.
You need these to generate your Wikipedia page.

Then we have "Servers.sh" and "Virtualservers.sh"

These 2 scripts are nearly the same, but they have a few differences.

Servers.sh collects data from Physical Machines. 

The only differences that these 2 scripts are:

Servers.sh uses a different method to check your physical machines Max Memory Size. It also checks if the machine has Virtualization available.
Virtualservers.sh on the other hand has 2 different Port variables. It has commands to check machines Open Internal and External ports.
Server.sh saves the data to "/tmp"phy_machines.txt" 
Virtualservers.sh saves the data to "/tmp/vm_machines.txt"


**proginst.sh**
This script is used to check if the following Physical Machines have the required programs installed. If the machine does not have any of the required programs installed, they will be installed.

**ports.sh**
This script is a an alternative to the "nmap" command which is used in "servers.sh" and "virtualservers.sh" to find open External Ports.
This script itself is not used, but i decided to keep it so the user has an option to choose from.
By Default, nmap is used as it is more reliable.

**extports.sh**
The following script is also not used in this project. This is the "Command Script" for "ports.sh".
Will decide later if these 2 scripts will be removed completely or not as nmap should give enough information to users.


**memmaht.sh and maht.sh**
maht.sh is used in "servers.sh" to find the machines Total Disk Size.  It Adds up all the disks using "lshw" and gives you the size in "GB"
memmaht.sh is used in "servers.sh" to find the machines Max Memory Size. "lshw" is also used in this.
These 2 scripts can only be used in Physical Machines as virtual machines use virtual disks and virtual memory.

**delete.sh**
This script is executed the last, once all of the other scripts have ran.
This uses "DeleteFiles" and "DeleteFiles1" functions. One is used for Physical Machines and the other is used for Virtual Machines.
This script deletes the files that were generated by the previous scripts mentioned and also the files that got sent to each Physical and Virtual Machines.



Things to keep in mind:

Right now, the user has to manually change a few lines if they have added new virtual machine to one of the physical machine.

Row 60 in "files.sh is **"printf ' \n|/9.vmfarm1 ' >>  vm_machines.txt"**
vmfarm1 is a Physical Machine that is currently hosting 9 virtual machines. If a new virtual machine is added to the vmfarm1,
the user has to change the number "9" to "10" and so on.

This is on DO TO list, to make it easier for the user, so they don't have to change anything in "files.sh".